[GGS Branch] Fixed randomsequence.randomdouble generating doubles in incorrect range.

+++ branches/ggs/src/GeoGen/random/RandomSequence.hpp
@@ -18,7 +18,7 @@ namespace geogen
 			inline int NextInt() { return this->mtRand(); }
 			inline int NextInt(int min, int max) { return min + ((unsigned)this->NextInt() % (int)(max - min + 1)); }
 			inline Height NextHeight() { return (Height)this->NextInt(HEIGHT_MIN, HEIGHT_MAX); }
-			inline double NextDouble() { return (double)(this->NextInt()) * (1. / 4294967295.); }
+			inline double NextDouble() { return (double)((unsigned)this->NextInt()) * (1. / 4294967295.); }
 		};
 	}
 }
